% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{eta}
\alias{eta}
\title{Calculating Squared Drift Parameter}
\usage{
eta(alpha, beta, ntest, type)
}
\arguments{
\item{alpha}{The desired type I error rate.}

\item{beta}{The desired type II error rate.}

\item{ntest}{The number of tests.}

\item{type}{The type of the alpha-spending function, k = 1 for O'Briend-Fleming's design and k = 2 for Pocock's design.}
}
\value{
The value of the drift parameter.
}
\description{
This function computes the squared drift parameter given the type I and II error rates,
the number of tests, and the alpha-spending function.
The code is revised from the algorithm/code by Lai (2013).
}
\examples{
powervec=c(0.519,0.532,0.559,0.581,0.601,0.622,0.606,0.646,0.661,0.683,0.723,0.721,0.709,
    0.758,0.753,0.764,0.827,0.782,0.805,0.806,0.836,0.85,0.85,0.854,0.877,0.888,0.8835,
    0.884,0.898,0.9175,0.9025,0.911,0.922,0.9265,0.936,0.935,0.9465,0.935,0.943,0.948)
sizevec = seq(from=300,to=1080,by=20)
alpha = 0.1
ntest = 3
type =  2

eta(alpha,1-powervec[1],ntest,type)

}
\references{
Reference: Lai D (2013) Sample size determination for group sequential test under fractional Brownian motion. Electronic J of Stat 7, 1957–1967.
}
